requirements.txt
python-dotenv==1.0.1
pandas==2.3.3
numpy==2.3.0
ccxt==4.3.92
SQLAlchemy==2.0.36
scikit-learn==1.5.2
plotly==5.24.1
streamlit==1.38.0
pydantic==2.9.2